+++ quill/include/quill/bundled/fmt/core.h
@@ -18,7 +18,7 @@
 #include <vector>
 
 // The fmt library version in the form major * 10000 + minor * 100 + patch.
-#define FMT_VERSION 60200
+#define FMT_VERSION 60201
 
 #ifdef __has_feature
 #  define FMT_HAS_FEATURE(x) __has_feature(x)
@@ -827,10 +827,12 @@ template <typename Char> struct string_value {
   std::size_t size;
 };
 
-template <typename Context> struct custom_value {
+template <typename Context>
+struct custom_value
+{
   using parse_context = basic_format_parse_context<typename Context::char_type>;
   const void* value;
-  void (*format)(const void* arg, parse_context& parse_ctx, Context& ctx);
+  void (*format)(const void* arg, typename Context::parse_context_type& parse_ctx, Context& ctx);
 };
 
 // A formatting argument value.
@@ -888,15 +890,14 @@ template <typename Context> class value {
   value(const named_arg_base<char_type>& val) { named_arg = &val; }
 
  private:
-  // Formats an argument of a custom type, such as a user-defined class.
-  template <typename T, typename Formatter>
-  static void format_custom_arg(
-      const void* arg, basic_format_parse_context<char_type>& parse_ctx,
-      Context& ctx) {
-    Formatter f;
-    parse_ctx.advance_to(f.parse(parse_ctx));
-    ctx.advance_to(f.format(*static_cast<const T*>(arg), ctx));
-  }
+   // Formats an argument of a custom type, such as a user-defined class.
+   template <typename T, typename Formatter>
+   static void format_custom_arg(const void* arg, typename Context::parse_context_type& parse_ctx, Context& ctx)
+   {
+     Formatter f;
+     parse_ctx.advance_to(f.parse(parse_ctx));
+     ctx.advance_to(f.format(*static_cast<const T*>(arg), ctx));
+   }
 };
 
 template <typename Context, typename T>
@@ -1061,12 +1062,12 @@ template <typename Context> class basic_format_arg {
    public:
     explicit handle(internal::custom_value<Context> custom) : custom_(custom) {}
 
-    void format(basic_format_parse_context<char_type>& parse_ctx,
-                Context& ctx) const {
+    void format(typename Context::parse_context_type& parse_ctx, Context& ctx) const
+    {
       custom_.format(custom_.value, parse_ctx, ctx);
     }
 
-   private:
+  private:
     internal::custom_value<Context> custom_;
   };
 
@@ -1207,15 +1208,18 @@ FMT_CONSTEXPR basic_format_arg<Context> make_arg(const T& value) {
   return arg;
 }
 
-template <bool IS_PACKED, typename Context, typename T,
-          FMT_ENABLE_IF(IS_PACKED)>
-inline value<Context> make_arg(const T& val) {
+// The type template parameter is there to avoid an ODR violation when using
+// a fallback formatter in one translation unit and an implicit conversion in
+// another (not recommended).
+template <bool IS_PACKED, typename Context, type, typename T, FMT_ENABLE_IF(IS_PACKED)>
+inline value<Context> make_arg(const T& val)
+{
   return arg_mapper<Context>().map(val);
 }
 
-template <bool IS_PACKED, typename Context, typename T,
-          FMT_ENABLE_IF(!IS_PACKED)>
-inline basic_format_arg<Context> make_arg(const T& value) {
+template <bool IS_PACKED, typename Context, type, typename T, FMT_ENABLE_IF(!IS_PACKED)>
+inline basic_format_arg<Context> make_arg(const T& value)
+{
   return make_arg<Context>(value);
 }
 
@@ -1272,7 +1276,9 @@ template <typename OutputIt, typename Char> class basic_format_context {
  public:
   using iterator = OutputIt;
   using format_arg = basic_format_arg<basic_format_context>;
-  template <typename T> using formatter_type = formatter<T, char_type>;
+  using parse_context_type = basic_format_parse_context<Char>;
+  template <typename T>
+  using formatter_type = formatter<T, char_type>;
 
   basic_format_context(const basic_format_context&) = delete;
   void operator=(const basic_format_context&) = delete;
@@ -1342,11 +1348,12 @@ class format_arg_store
                 : internal::is_unpacked_bit | num_args;
 
   format_arg_store(const Args&... args)
-      :
+    :
 #if FMT_GCC_VERSION && FMT_GCC_VERSION < 409
-        basic_format_args<Context>(*this),
+      basic_format_args<Context>(*this),
 #endif
-        data_{internal::make_arg<is_packed, Context>(args)...} {
+      data_{internal::make_arg<is_packed, Context, internal::mapped_type_constant<Args, Context>::value>(args)...}
+  {
   }
 };